pyproject.toml
# Documentation for pyproject.toml: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

# name and version are mandatory, change the values but do not delete them.
[project]
name = "comp0034-cw2"
version = "2025.1"
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-WTF",
    "pandas",
    "plotly",
    "pytest",
    "pytest-cov"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "error",
]

[tool.coverage.run]
branch = true
source = ["flaskstarter", "tests"]

[tool.coverage.paths]
source = ["src", "flaskstarter"]